Refactoring
extract out Mail classes to core
